feat: remove 1Gwei minimum fee requirement

+++ scenarios/uniswap-swaps/uniswap_swaps.go
@@ -23,22 +23,22 @@ import (
 )
 
 type ScenarioOptions struct {
-	TotalCount    uint64 `yaml:"total_count"`
-	Throughput    uint64 `yaml:"throughput"`
-	MaxPending    uint64 `yaml:"max_pending"`
-	MaxWallets    uint64 `yaml:"max_wallets"`
-	Rebroadcast   uint64 `yaml:"rebroadcast"`
-	BaseFee       uint64 `yaml:"base_fee"`
-	TipFee        uint64 `yaml:"tip_fee"`
-	PairCount     uint64 `yaml:"pair_count"`
-	MinSwapAmount string `yaml:"min_swap_amount"`
-	MaxSwapAmount string `yaml:"max_swap_amount"`
-	BuyRatio      uint64 `yaml:"buy_ratio"`
-	Slippage      uint64 `yaml:"slippage"`
-	SellThreshold string `yaml:"sell_threshold"`
-	Timeout       string `yaml:"timeout"`
-	ClientGroup   string `yaml:"client_group"`
-	LogTxs        bool   `yaml:"log_txs"`
+	TotalCount    uint64  `yaml:"total_count"`
+	Throughput    uint64  `yaml:"throughput"`
+	MaxPending    uint64  `yaml:"max_pending"`
+	MaxWallets    uint64  `yaml:"max_wallets"`
+	Rebroadcast   uint64  `yaml:"rebroadcast"`
+	BaseFee       float64 `yaml:"base_fee"`
+	TipFee        float64 `yaml:"tip_fee"`
+	PairCount     uint64  `yaml:"pair_count"`
+	MinSwapAmount string  `yaml:"min_swap_amount"`
+	MaxSwapAmount string  `yaml:"max_swap_amount"`
+	BuyRatio      uint64  `yaml:"buy_ratio"`
+	Slippage      uint64  `yaml:"slippage"`
+	SellThreshold string  `yaml:"sell_threshold"`
+	Timeout       string  `yaml:"timeout"`
+	ClientGroup   string  `yaml:"client_group"`
+	LogTxs        bool    `yaml:"log_txs"`
 }
 
 type Scenario struct {
@@ -89,8 +89,8 @@ func (s *Scenario) Flags(flags *pflag.FlagSet) error {
 	flags.Uint64Var(&s.options.MaxPending, "max-pending", ScenarioDefaultOptions.MaxPending, "Maximum number of pending transactions")
 	flags.Uint64Var(&s.options.MaxWallets, "max-wallets", ScenarioDefaultOptions.MaxWallets, "Maximum number of child wallets to use")
 	flags.Uint64Var(&s.options.Rebroadcast, "rebroadcast", ScenarioDefaultOptions.Rebroadcast, "Enable reliable rebroadcast system")
-	flags.Uint64Var(&s.options.BaseFee, "basefee", ScenarioDefaultOptions.BaseFee, "Max fee per gas to use in transfer transactions (in gwei)")
-	flags.Uint64Var(&s.options.TipFee, "tipfee", ScenarioDefaultOptions.TipFee, "Max tip per gas to use in transfer transactions (in gwei)")
+	flags.Float64Var(&s.options.BaseFee, "basefee", ScenarioDefaultOptions.BaseFee, "Max fee per gas to use in transfer transactions (in gwei)")
+	flags.Float64Var(&s.options.TipFee, "tipfee", ScenarioDefaultOptions.TipFee, "Max tip per gas to use in transfer transactions (in gwei)")
 	flags.Uint64Var(&s.options.PairCount, "pair-count", ScenarioDefaultOptions.PairCount, "Number of uniswap pairs to deploy")
 	flags.StringVar(&s.options.MinSwapAmount, "min-swap", ScenarioDefaultOptions.MinSwapAmount, "Minimum swap amount in wei")
 	flags.StringVar(&s.options.MaxSwapAmount, "max-swap", ScenarioDefaultOptions.MaxSwapAmount, "Maximum swap amount in wei")
